+++ actix-rt/CHANGES.md
@@ -1,6 +1,9 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Add `Arbiter::try_current` for situations where thread may or may not have Arbiter context. [#408]
+
+[#408]: https://github.com/actix/actix-net/pull/408
 
 
 ## 2.3.0 - 2021-10-11